refactor: 폴더 구조를 Go Clean Architecture 기반으로 변경
