+++ cmd/infrastructure/repository/video_repository_impl.go
@@ -1,21 +1,16 @@
 package repository
 
 import (
-	"github.com/HongJungWan/ffmpeg-video-modules/app/domain"
+	"github.com/HongJungWan/ffmpeg-video-modules/cmd/domain"
+	"github.com/HongJungWan/ffmpeg-video-modules/cmd/domain/repository"
 	"gorm.io/gorm"
 )
-
-type VideoRepository interface {
-	Save(video *domain.Video) error
-	FindByID(videoID int) (*domain.Video, error)
-	FindAll() ([]domain.Video, error)
-}
 
 type VideoRepositoryImpl struct {
 	DB *gorm.DB
 }
 
-func NewVideoRepository(db *gorm.DB) *VideoRepositoryImpl {
+func NewVideoRepository(db *gorm.DB) repository.VideoRepository {
 	return &VideoRepositoryImpl{DB: db}
 }
