+++ cmd/infrastructure/router/router.go
@@ -1,11 +1,11 @@
 package router
 
 import (
-	"github.com/HongJungWan/ffmpeg-video-modules/app/domain"
-	"github.com/HongJungWan/ffmpeg-video-modules/app/infrastructure/configs"
-	"github.com/HongJungWan/ffmpeg-video-modules/app/interfaces/controller"
-	"github.com/HongJungWan/ffmpeg-video-modules/app/interfaces/repository"
-	"github.com/HongJungWan/ffmpeg-video-modules/app/usecases"
+	"github.com/HongJungWan/ffmpeg-video-modules/cmd/domain"
+	"github.com/HongJungWan/ffmpeg-video-modules/cmd/infrastructure/configs"
+	repository2 "github.com/HongJungWan/ffmpeg-video-modules/cmd/infrastructure/repository"
+	"github.com/HongJungWan/ffmpeg-video-modules/cmd/interfaces/controller"
+	"github.com/HongJungWan/ffmpeg-video-modules/cmd/usecases"
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
@@ -38,9 +38,9 @@ func NewRouter(conf configs.Config, db *gorm.DB) *gin.Engine {
 	healthCheckController := controller.NewHealthCheckController(healthCheckInteractor)
 
 	// 비즈니스 로직 관련 설정
-	videoRepository := repository.NewVideoRepository(db)
-	videoJobRepository := repository.NewVideoJobRepository(db)
-	finalVideoRepository := repository.NewFinalVideoRepository(db)
+	videoRepository := repository2.NewVideoRepository(db)
+	videoJobRepository := repository2.NewVideoJobRepository(db)
+	finalVideoRepository := repository2.NewFinalVideoRepository(db)
 
 	videoInteractor := usecases.NewVideoInteractor(videoRepository, videoJobRepository, finalVideoRepository)
 	videoJobInteractor := usecases.NewVideoJobInteractor(videoJobRepository, videoRepository, finalVideoRepository)
