+++ cmd/interfaces/dto/request/trim_request.go
