+++ cmd/usecases/video_interactor.go
@@ -2,10 +2,10 @@ package usecases
 
 import (
 	"fmt"
-	"github.com/HongJungWan/ffmpeg-video-modules/app/domain"
-	"github.com/HongJungWan/ffmpeg-video-modules/app/infrastructure/ffmpeg"
-	"github.com/HongJungWan/ffmpeg-video-modules/app/interfaces/dto/response"
-	"github.com/HongJungWan/ffmpeg-video-modules/app/interfaces/repository"
+	"github.com/HongJungWan/ffmpeg-video-modules/cmd/domain"
+	"github.com/HongJungWan/ffmpeg-video-modules/cmd/domain/repository"
+	"github.com/HongJungWan/ffmpeg-video-modules/cmd/infrastructure/ffmpeg"
+	"github.com/HongJungWan/ffmpeg-video-modules/cmd/interfaces/dto/response"
 	"github.com/gin-gonic/gin"
 	"log"
 	"net/http"
